Added generic request handlers
The generic request handler would allow new resource types to be easily fit
into either warm resources i.e resources present in warm pool before the incoming
request comes in and on demand resources i.e resources that are created after
receiving the incoming request. If in future we come up with a different kind of resource
that cannot fit these two categories we can easily introduce a new handler for it.