Added the node reconciler module
The node reconciler module consists of the reconciler which listens
for the node events and passes the events to the node manager. The
node manager maintains the list of node that it manages. The new nodes
are added to the list if the node is eligible for management and
all the supported resources are initialized. In case of delete event
the node is removed from the managed list and resources are cleaned up.
The node manager supports concurrent operations so it's safe to increase
the number of worker routines of the node reconciler.