Added custom controller to store stripped down pod details
We are adding a custom controller in order to reduce the memory usage
of the container. The custom controller will allow us to store only
required details of k8s object and strip away the unnecessary info to
help bring down memory significantly for large clusters.

The custom controller does the following
- Watch: In watch operation the k8s object details are stripped away
  in the ProcessFunc before it's added to the Data Store
- List: In the list operation we do pagination and intercept the
  pagination response before it reaches the caller. On intercepting
  the page of k8s object we call our converter to again strip down the
  k8s object details.
- Kube builder Controller Integration: To not lose the ability to do
  timed retries we notify the kube builder controller of the pod events
  using a channel where we pass Generic Events in ProcessFunc.